[flatland] Fix all implicit conversions in flatland/
Fixes all implicit conversions in the flatland/ directory
and turns compiler warnings back on.

Bug: 58162
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779966
Commit-Queue: Mikael Pessa <mikaelpessa@google.com>

+++ src/ui/scenic/lib/display/util.h
@@ -38,7 +38,7 @@ bool IsCaptureSupported(const fuchsia::hardware::display::ControllerSyncPtr& dis
 uint64_t ImportImageForCapture(
     const fuchsia::hardware::display::ControllerSyncPtr& display_controller,
     const fuchsia::hardware::display::ImageConfig& image_config,
-    allocation::GlobalBufferCollectionId buffer_collection_id, uint64_t vmo_idx);
+    allocation::GlobalBufferCollectionId buffer_collection_id, uint32_t vmo_idx);
 
 }  // namespace scenic_impl
